Add built-in first plerson player model
